+++ src/Libraries/SmartStore.Data/Migrations/MigrationsConfiguration.cs
@@ -38,6 +38,15 @@
 			// SeoSettings.RedirectLegacyTopicUrls should be true when migrating (it is false by default after fresh install)
 			var name = TypeHelper.NameOf<SeoSettings>(y => y.RedirectLegacyTopicUrls, true);
 			context.MigrateSettings(x => x.Add(name, true));
+
+			// remove setting which were moved from customer settings to privacy settings and have new default values which should be applied immediately 
+			var settings = context.Set<Setting>();
+			var storeLastIpAddressSetting = settings.FirstOrDefault(x => x.Name == "CustomerSettings.StoreLastIpAddress");
+			if (storeLastIpAddressSetting != null) settings.Remove(storeLastIpAddressSetting);
+			
+			var displayPrivacyAgreementOnContactUs = settings.FirstOrDefault(x => x.Name == "CustomerSettings.DisplayPrivacyAgreementOnContactUs");
+			if (displayPrivacyAgreementOnContactUs != null) settings.Remove(displayPrivacyAgreementOnContactUs);
+
 		}
 
 		public void MigrateLocaleResources(LocaleResourcesBuilder builder)
@@ -65,7 +74,24 @@
 				"{0} benutzt Cookies, um Ihnen das beste Einkaufs-Erlebnis zu ermöglichen. Zum Teil werden Cookies auch von Drittanbietern gesetzt. <a href='{1}'>Datenschutzerklärung</a>");
 
 			builder.AddOrUpdate("CookieConsent.Button", "Okay, got it", "Ok, verstanden");
-			
+
+			builder.Delete("ContactUs.PrivacyAgreement");
+
+			builder.Delete("Admin.Configuration.Settings.CustomerUser.StoreLastIpAddress");
+			builder.Delete("Admin.Configuration.Settings.CustomerUser.StoreLastIpAddress.Hint");
+			builder.AddOrUpdate("Admin.Configuration.Settings.CustomerUser.Privacy.StoreLastIpAddress",
+				"Store IP address",
+				"IP-Adresse speichern",
+				"Specifies whether to store the IP address in the customer data set.",
+				"Legt fest, ob die IP-Adresse im Kundendatensatz gespeichert werden soll.");
+
+			builder.Delete("Admin.Configuration.Settings.CustomerUser.DisplayPrivacyAgreementOnContactUs");
+			builder.Delete("Admin.Configuration.Settings.CustomerUser.DisplayPrivacyAgreementOnContactUs.Hint");
+			builder.AddOrUpdate("Admin.Configuration.Settings.CustomerUser.Privacy.DisplayPrivacyAgreementOnContactUs",
+				"Get privacy consent for contact requests",
+				"Einwilligungserklärung im Kontaktformular fordern",
+				"Specifies whether a checkbox will be displayed on the contact page which requests the user to agree on storage of his data.",
+				"Bestimmt ob im Kontaktformular eine Checkbox angezeigt wird, die den Benutzer auffordert der Speicherung seiner Daten zuzustimmen.");
 		}
 	}
 }