+++ src/Presentation/SmartStore.Web/Controllers/HomeController.cs
@@ -36,6 +36,7 @@ namespace SmartStore.Web.Controllers
 		private readonly Lazy<CommonSettings> _commonSettings;
 		private readonly Lazy<SeoSettings> _seoSettings;
 		private readonly Lazy<CustomerSettings> _customerSettings;
+		private readonly Lazy<PrivacySettings> _privacySettings;
 
 		public HomeController(
 			ICommonServices services,
@@ -49,7 +50,8 @@ namespace SmartStore.Web.Controllers
 			Lazy<CaptchaSettings> captchaSettings,
 			Lazy<CommonSettings> commonSettings,
 			Lazy<SeoSettings> seoSettings,
-			Lazy<CustomerSettings> customerSettings)
+			Lazy<CustomerSettings> customerSettings,
+			Lazy<PrivacySettings> privacySettings)
         {
 			this._services = services;
 			this._categoryService = categoryService;
@@ -63,7 +65,8 @@ namespace SmartStore.Web.Controllers
 			this._commonSettings = commonSettings;
 			this._seoSettings = seoSettings;
             this._customerSettings = customerSettings;
-        }
+			this._privacySettings = privacySettings;
+		}
 
 
         [RequireHttpsByConfigAttribute(SslRequirement.No)]
@@ -87,7 +90,7 @@ namespace SmartStore.Web.Controllers
 				Email = _services.WorkContext.CurrentCustomer.Email,
 				FullName = _services.WorkContext.CurrentCustomer.GetFullName(),
 				DisplayCaptcha = _captchaSettings.Value.Enabled && _captchaSettings.Value.ShowOnContactUsPage,
-                DisplayPrivacyAgreement = _customerSettings.Value.DisplayPrivacyAgreementOnContactUs,
+                DisplayPrivacyAgreement = _privacySettings.Value.DisplayPrivacyAgreementOnContactUs,
                 MetaKeywords = topic?.GetLocalized(x => x.MetaKeywords),
                 MetaDescription = topic?.GetLocalized(x => x.MetaDescription),
                 MetaTitle = topic?.GetLocalized(x => x.MetaTitle),